feat(custom): 直播间分页查询性能优化，新增直播间开播信息校验接口

+++ live-common-interface/src/main/java/top/flobby/live/common/req/PageBaseReq.java
@@ -12,6 +12,6 @@ import lombok.Data;
 @Data
 public class PageBaseReq {
 
-    private int page;
-    private int pageSize;
+    private Long page;
+    private Long pageSize;
 }
